requirements.txt
requests>=2.32.5
pandas>=2.3.3
openpyxl>=3.1.5
ollama>=0.6.0
python-dotenv>=1.2.1

